Canonicalize aux build paths relative to the current directory.
Otherwise, within the rustc repo, we get an aux build path relative to the repository root, which we then can't resolve to the actual aux file
